+++ fish-report/src/main/java/com/yihuacomputer/fish/report/service/trans/etl/TransTypeEtlService.java
@@ -99,12 +99,12 @@ public class TransTypeEtlService implements ITransTypeEtlService{
 	}
 
 	@Override
-	public List<ITransTypeWeek> getWeek(int weekOfYear) {
+	public List<ITransTypeWeek> getWeek(long weekOfYear) {
 		return dao.findByHQL("from TransTypeWeek where date = ?", weekOfYear);
 	}
 
 	@Override
-	public Long[] getWeekTotal(int weekOfYear) {
+	public Long[] getWeekTotal(long weekOfYear) {
 		StringBuilder sql = new StringBuilder();
 		sql.append("select sum(rc.TRANS_COUNT),sum(rc.TRANS_AMOUNT)");
 		sql.append(" from etl_trans_type_week rc");
@@ -125,12 +125,12 @@ public class TransTypeEtlService implements ITransTypeEtlService{
 	}
 
 	@Override
-	public List<ITransTypeMonth> getMonth(int month) {
+	public List<ITransTypeMonth> getMonth(long month) {
 		return dao.findByHQL("from TransTypeMonth where date = ?", month);
 	}
 
 	@Override
-	public Long[] getMonthTotal(int month) {
+	public Long[] getMonthTotal(long month) {
 		StringBuilder sql = new StringBuilder();
 		sql.append("select sum(rc.TRANS_COUNT),sum(rc.TRANS_AMOUNT)");
 		sql.append(" from etl_trans_type_month rc");
